+++ engines/bladerunner/script/ai/generic_walker_b.cpp
@@ -375,7 +375,7 @@ void AIScriptGenericWalkerB::movingStart() {
 	deltaX = walkerX - mccoyX;
 	deltaZ = walkerZ - mccoyZ;
 
-	float dist = sqrt(deltaX * deltaX + deltaZ * deltaZ);
+	float dist = sqrtf(deltaX * deltaX + deltaZ * deltaZ);
 	if (dist == 0.0f) {
 		deltaZ = 0.0f;
 		deltaX = 0.0f;
